requirements.txt
langchain==0.3.26
flask==3.1.1
sentence-transformers==3.3.1
transformers==4.46.2
torch>=2.2.0
tensorflow==2.19.1
tf-keras==2.19.0
protobuf<5
pypdf==5.6.1
python-dotenv==1.1.0
langchain-pinecone==0.2.8
langchain-openai==0.3.24
langchain-community==0.3.26
-e .